add even more minimalistic example of jinja templating
